refactor: merge type decls about relays into relay.ts
