+++ packages/nostr-fetch/src/relay.ts
@@ -1,17 +1,8 @@
-import type {
-  RelayEventCbTypes,
-  RelayEventTypes,
-  RelayOptions,
-  SubEventCbTypes,
-  SubEventTypes,
-  Subscription,
-  SubscriptionOptions,
-  WSCloseEvent,
-} from "./relayTypes";
-
 import { verifyEventSig } from "@nostr-fetch/kernel/crypto";
 import type { C2RMessage, Filter, NostrEvent } from "@nostr-fetch/kernel/nostr";
 import { generateSubId, parseR2CMessage } from "@nostr-fetch/kernel/nostr";
+
+type Callback<E> = E extends void ? () => void : (ev: E) => void;
 
 export interface Relay {
   url: string;
@@ -25,9 +16,33 @@ export interface Relay {
   off<E extends RelayEventTypes>(type: E, cb: RelayEventCbTypes[E]): void;
 }
 
+export type RelayOptions = {
+  connectTimeoutMs: number;
+};
+
 export const initRelay = (relayUrl: string, options: RelayOptions): Relay => {
   return new RelayImpl(relayUrl, options);
 };
+
+export type WSCloseEvent = {
+  code: number;
+  reason: string;
+  wasClean: boolean | undefined; // optional since websocket-polyfill's CloseEvent doesn't have it
+};
+
+export type RelayConnectCb = Callback<void>;
+export type RelayDisconnectCb = Callback<WSCloseEvent | undefined>;
+export type RelayNoticeCb = Callback<unknown>;
+export type RelayErrorCb = Callback<void>;
+
+export type RelayEventCbTypes = {
+  connect: RelayConnectCb;
+  disconnect: RelayDisconnectCb;
+  notice: RelayNoticeCb;
+  error: RelayErrorCb;
+};
+
+export type RelayEventTypes = keyof RelayEventCbTypes;
 
 type RelayListenersTable = {
   [E in RelayEventTypes]: Set<RelayEventCbTypes[E]>;
@@ -189,6 +204,34 @@ class RelayImpl implements Relay {
     }
     this.#ws.send(JSON.stringify(msg));
   }
+}
+
+type EoseEventPayload = {
+  aborted: boolean;
+};
+
+export type SubEventCb = Callback<NostrEvent>;
+export type SubEoseCb = Callback<EoseEventPayload>;
+
+export type SubEventCbTypes = {
+  event: SubEventCb;
+  eose: SubEoseCb;
+};
+
+export type SubEventTypes = keyof SubEventCbTypes;
+
+export interface Subscription {
+  subId: string;
+  req(): void;
+  close(): void;
+  on<E extends SubEventTypes>(type: E, cb: SubEventCbTypes[E]): void;
+  off<E extends SubEventTypes>(type: E, cb: SubEventCbTypes[E]): void;
+}
+
+export interface SubscriptionOptions {
+  subId?: string;
+  skipVerification: boolean;
+  abortSubBeforeEoseTimeoutMs: number;
 }
 
 type SubListenersTable = {
